Add incremental behavior replication
Concretely this introduced the ability for the runtime to 'reset'
client values that depend on server values. For incremental behaviors
concretely this means that changes of type A can reset Inc[A, DeltaA]'s
values without providing DeltaA's. This gives us the recover ability with
Server sent events aggressive reconnecting.